[TodoApp] added implementation

+++ src/main/java/be/app/Todo/service/TaskServiceImpl.java
@@ -2,10 +2,13 @@ package be.app.Todo.service;
 
 import be.app.Todo.dao.TaskDAO;
 import be.app.Todo.entity.Task;
+import be.app.Todo.exception.NotFoundException;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -15,8 +18,12 @@ public class TaskServiceImpl implements TaskService {
     TaskDAO taskDAO;
 
     @Override
-    public Task getTask(String title) {
-        return taskDAO.getTask(title);
+    public Task getTask(Long id) {
+        Task task= taskDAO.getTask(id);
+        if(task==null){
+            throw new NotFoundException("Task not found id :" + id);
+        }
+        return task;
     }
 
     @Override
@@ -25,22 +32,37 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
+    @Transactional
     public Task addTask(Task task) {
-        System.out.println(task.getDueDate());
-        task.setIsDeleted(0);
-        LocalDate localDate = LocalDate.now();
-        task.setCreated(localDate);
+        task.setIsDeleted(false);
+        LocalDateTime localDateTime= LocalDateTime.now();
+        task.setCreated(localDateTime);
         return taskDAO.addTask(task);
     }
 
     @Override
-    public Task editTask(Task task) {
+    @Transactional
+    public Task editTask(Task task, Long id) {
+        Task oldTask = taskDAO.getTask(id);
+        if(oldTask==null){
+            throw new NotFoundException("Task not found id :" + task.getTitle());
+        }
+        task.setId(oldTask.getId());
+        task.setCreated(oldTask.getCreated());
+        LocalDateTime localDateTime= LocalDateTime.now();
+        task.setUpdated(localDateTime);
+        task.setIsDeleted(oldTask.getIsDeleted());
         return taskDAO.editTask(task);
     }
 
     @Override
-    public Task deleteTask(Task task) {
-        task.setIsDeleted(1);
+    @Transactional
+    public Task deleteTask(Long id) {
+        Task task = taskDAO.getTask(id);
+        if(task==null){
+            throw new NotFoundException("Task not found id :" + id);
+        }
+        task.setIsDeleted(true);
         return taskDAO.deleteTask(task);
     }
 }
